Splitting apart vision sensors, adding action groups to expert sensor
