+++ projects/objectnav_baselines/experiments/objectnav_mixin_resnetgru.py
@@ -5,7 +5,7 @@ import torch.nn as nn
 from torchvision import models
 
 from allenact.base_abstractions.preprocessor import Preprocessor
-from allenact.base_abstractions.sensor import RGBSensor, DepthSensor
+from allenact.embodiedai.sensors.vision_sensors import RGBSensor, DepthSensor
 from allenact.embodiedai.preprocessors.resnet import ResNetPreprocessor
 from allenact.utils.experiment_utils import Builder
 from allenact_plugins.ithor_plugin.ithor_sensors import GoalObjectTypeThorSensor
